adiciina auth and autor

+++ src/main/java/br/letscode/ecommerce/service/implementation/ClienteServiceImpl.java
@@ -4,13 +4,16 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import br.letscode.ecommerce.dao.ClienteDao;
-import br.letscode.ecommerce.dto.ClienteDto;
+import br.letscode.ecommerce.dto.ClienteDtoComSenha;
+import br.letscode.ecommerce.dto.ClienteDtoSemSenha;
 import br.letscode.ecommerce.entity.Cliente;
 import br.letscode.ecommerce.exceptions.EcommerceException;
 import br.letscode.ecommerce.models.Message;
+import br.letscode.ecommerce.models.PerfilEnum;
 import br.letscode.ecommerce.service.ClienteService;
 
 @Service
@@ -19,36 +22,49 @@ public class ClienteServiceImpl implements ClienteService {
   @Autowired
   ClienteDao clienteDao;
 
+  @Autowired
+  private final PasswordEncoder passwordEncoder;
+
+  public ClienteServiceImpl(PasswordEncoder passwordEncoder) {
+    this.passwordEncoder = passwordEncoder;
+  }
+
   @Override
   public List<Cliente> getAllClients() {
     return clienteDao.findAll();
   }
 
   @Override
-  public Message novoCliente(ClienteDto clienteDto) {
+  public Message novoCliente(ClienteDtoComSenha clienteDto) {
+    clienteDto.setSenha(passwordEncoder.encode(clienteDto.getSenha()));
     Cliente cliente = new Cliente(clienteDto.getNome(), clienteDto.getSobrenome(), clienteDto.getEmail(),
-        clienteDto.getSexo(), clienteDto.getCpf());
+        clienteDto.getSexo(), clienteDto.getCpf(), clienteDto.getSenha(), PerfilEnum.CLIENTE);
     clienteDao.save(cliente);
+    cliente.setSenha("");
     return new Message("Cliente: " + cliente + " salvo");
   }
 
   @Override
-  public Message atualizarCliente(ClienteDto clienteDto, Long id) {
+  public Message atualizarCliente(ClienteDtoSemSenha clienteDto, Long id) {
     Optional<Cliente> clienteFoundById = clienteDao.findById(id);
     if (clienteFoundById.isEmpty()) {
       throw new EcommerceException("Cliente não existe", 404);
     }
     Cliente cliente = new Cliente(clienteFoundById.get().getId(), clienteDto.getNome(), clienteDto.getSobrenome(),
         clienteDto.getEmail(),
-        clienteDto.getSexo(), clienteDto.getCpf());
+        clienteDto.getSexo(), clienteDto.getCpf(), clienteFoundById.get().getSenha(), clienteDto.getPerfilEnum()  );
     clienteDao.save(cliente);
     return new Message("Cliente: " + cliente + " salvo com sucesso");
   }
 
   @Override
-  public Message removerCliente(long id) {
+  public Message removerCliente(Long id) {
+    Optional<Cliente> clienteOptional = clienteDao.findById(id);
+    if (clienteOptional.isEmpty()) {
+      throw new EcommerceException("Cliente não existe", 404);
+    }
     clienteDao.deleteById(id);
-    return new Message("Cliente: " + clienteDao.findById(id) + " deletado com sucesso");
+    return new Message("Cliente: " + clienteOptional.get() + " deletado com sucesso");
   }
 
   @Override
@@ -57,5 +73,14 @@ public class ClienteServiceImpl implements ClienteService {
       throw new EcommerceException("Cliente não existe", 404);
     }
     return clienteDao.findByCpf(cpf).get();
+  }
+
+  @Override
+  public Cliente findById(Long id) {
+    Optional<Cliente> clienteOptional = clienteDao.findById(id);
+    if (clienteOptional.isEmpty()) {
+      throw new EcommerceException("Cliente não existe", 404);
+    }
+    return clienteOptional.get();
   }
 }
